% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{directions}
\alias{directions}
\title{Directions Function}
\usage{
directions(from, to)
}
\arguments{
\item{from}{The starting location of your designated route. Must be a string in the format: "Street Address, City, State"}

\item{to}{The ending location of your designated route. Must be a string in the format: "Street Address, City, State"}
}
\value{
List of Data Table, Leaflet Map
}
\description{
A function to provide visual and textual directions from a starting location to an ending location.
}
