% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{get_query}
\alias{get_query}
\title{Helper function to query the data from the MapQuest API}
\usage{
get_query(from, to)
}
\arguments{
\item{from}{The starting location of your designated route. Must be a string in the format: "Street Address, City, State"}

\item{to}{The ending location of your designated route. Must be a string in the format: "Street Address, City, State"}
}
\value{
A List of data from query specifying directions, latitude and longitude points of each direction, etc.
}
\description{
Helper function to query the data from the MapQuest API
}
