% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{this function uses the list of lists from the returned element in query_weather and creates a dataframe to show all weather
requested in a nice manner and for the dates specified}
\usage{
make_output(big_query, num_days)
}
\description{
this function uses the list of lists from the returned element in query_weather and creates a dataframe to show all weather
requested in a nice manner and for the dates specified
}
