% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{query_weather}
\alias{query_weather}
\title{This is the function that calls the API with the given dates from the date_input fucntion and the location from the
directions function.}
\usage{
query_weather(fields, lat_input, lng_input, dates)
}
\arguments{
\item{fields}{this is a list of weather fields that you want to collect in your weather query from climacell,
the full list of offered fields are given at https://developer.climacell.co/v3/docs/forecast in the "Daily" section,
this should be in a list format}

\item{lat_input}{this is a latitude where you want to collect information on the given weather fields during the given
dates}

\item{lng_input}{this is a longitude where you want to collect information on the given weather fields during the given
dates}

\item{dates}{this is the output of the date_input function and gives the starting and ending date for the query to
collect weather dat for}
}
\value{
this returns a list of x lists that each include the output of the query for the particular weather field associated
with it (x is the number of fields you listed in the fields parameter)
}
\description{
This is the function that calls the API with the given dates from the date_input fucntion and the location from the
directions function.
}
