% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{date_input}
\alias{date_input}
\title{Date Input Helper Function}
\usage{
date_input(day_leaving, day_returning)
}
\arguments{
\item{day_leaving}{the date the user is planning to leave for their vacation in yyyy-mm-dd format}

\item{day_returning}{the last date the user is planning to be on their vacation in yyyy-mm-dd format}
}
\value{
a vector that includes the start time of the vacation and the end time in ISO8601 format for the API
}
\description{
This is a helper function for query_weather that takes in dates in yyyy-mm-dd format and outputs them in a form that the API can handle:
}
