% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Make Weather Forecasting Function}
\usage{
make_output(fields, lat_input, lng_input, day_leaving, day_returning, num_days)
}
\arguments{
\item{fields}{this is a list of weather fields that you want to collect in your weather query from climacell,
the full list of offered fields are given at https://developer.climacell.co/v3/docs/forecast in the "Daily" section,
this should be in a list format}

\item{lat_input}{this is a latitude where you want to collect information on the given weather fields during the given dates}

\item{lng_input}{this is a longitude where you want to collect information on the given weather fields during the given dates}

\item{day_leaving}{the date the user is planning to leave for their vacation in yyyy-mm-dd format}

\item{day_returning}{the last date the user is planning to be on their vacation in yyyy-mm-dd format}

\item{num_days}{this is the number of days between the start date and the end date of the vacation, this
should be calculated with dates\link{2} - dates\link{1} + 2}
}
\value{
a data frame with all weather for the given dates above and the given fields above
}
\description{
The make_output function greats a data frame that includes weather forecasting for up to 15 days from today, after a user inputs
a starting data and ending date of a trip. Th function uses another function in this package to get an address and plugs it into
a query for a weather API to collect the forecast on that location.
}
