% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{query_weather}
\alias{query_weather}
\title{Query Weather Helper Function}
\usage{
query_weather(fields, lat_input, lng_input, day_leaving, day_returning)
}
\arguments{
\item{fields}{this is a list of weather fields that you want to collect in your weather query from climacell,
the full list of offered fields are given at https://developer.climacell.co/v3/docs/forecast in the "Daily" section,
this should be in a list format}

\item{lat_input}{this is a latitude where you want to collect information on the given weather fields during the given dates. Must be numeric.}

\item{lng_input}{this is a longitude where you want to collect information on the given weather fields during the given dates. Must be numeric.}

\item{day_leaving}{the date the user is planning to leave for their vacation in yyyy-mm-dd format. Must be a string.}

\item{day_returning}{the last date the user is planning to be on their vacation in yyyy-mm-dd format. Must be a string.}
}
\value{
this returns a list of x lists that each include the output of the query for the particular weather field associated
with it (x is the number of fields you listed in the fields parameter)
}
\description{
This is the function that calls the API with the given dates from the date_input fucntion and the location from the
directions function.
}
